Add Navbar, Footer, and SearchBar components; update app layout and styles
