+++ frontend/components/Footer.tsx
@@ -1,0 +1,9 @@
+export default function Footer() {
+  return (
+    <footer className=" bg-[#3A7858] text-white/90">
+      <div className="mx-auto max-w-6xl px-6 py-10">
+        <div className="text-3xl font-extrabold">Rentify</div>
+      </div>
+    </footer>
+  );
+}
