+++ frontend/pages/_app.tsx
@@ -1,6 +1,19 @@
 import "@/styles/globals.css";
 import type { AppProps } from "next/app";
+import Navbar from "@/components/Navbar";
+import Footer from "@/components/Footer";
 
 export default function App({ Component, pageProps }: AppProps) {
-  return <Component {...pageProps} />;
+  return (
+    <div
+      className="min-h-screen flex flex-col"
+      style={{ backgroundColor: "#F4F6F5" }}
+    >
+      <Navbar />
+      <div className="flex-1">
+        <Component {...pageProps} />
+      </div>
+      <Footer />
+    </div>
+  );
 }
