+++ frontend/styles/globals.css
@@ -1,7 +1,7 @@
 @import "tailwindcss";
 
 :root {
-  --background: #ffffff;
+  --background: #f4f6f5;
   --foreground: #171717;
 }
 
@@ -19,8 +19,18 @@
   }
 }
 
+/* Remove default browser margins to avoid white gutters */
+html,
 body {
+  height: 100%;
+}
+body {
+  margin: 0;
   background: var(--background);
   color: var(--foreground);
   font-family: Arial, Helvetica, sans-serif;
+}
+
+nextjs-portal {
+  display: none !important;
 }
